medical_RAG_system/requirements.txt
pymupdf
tqdm
spacy
torch
sentence-transformers
elasticsearch
faiss-cpu
numpy
pandas
flask
streamlit
requests
python-dotenv
openai
google-generativeai
